refactor: Clean up repository structure
- Remove unnecessary development and test files
- Move all build scripts to build_scripts/ directory
- Update all references to moved files
- Update .gitignore for new structure
- Remove obsolete documentation files
