--- RTF_IMPLEMENTATION_SUMMARY.md
@@ -1,143 +1,0 @@
-# RTF Processing Implementation Summary
-
-## Overview
-
-This implementation provides comprehensive RTF parsing, section detection, and text cleaning for the EasyWorship to ProPresenter converter project. The modules are specifically designed to handle EasyWorship's RTF format with Swedish language support.
-
-## Implemented Modules
-
-### 1. RTF Parser (`src/processing/rtf_parser.py`)
-
-**Purpose**: Parse RTF content from EasyWorship's SongWords.db database with Swedish Unicode support.
-
-**Key Features**:
-- Utilizes `striprtf` library as the foundation with custom post-processing
-- Handles Swedish Unicode characters: `\u228?` → ä, `\u229?` → å, `\u246?` → ö, `\u180?` → ´
-- Converts RTF structure: `\par` → paragraph breaks, `\line` → line breaks
-- Fallback manual parsing when striprtf fails
-- Proper error handling and logging
-
-**Main Class**: `EasyWorshipRTFParser`
-**Convenience Function**: `parse_rtf(rtf_content)`
-
-### 2. Section Detector (`src/processing/section_detector.py`)
-
-**Purpose**: Detect and structure song sections from parsed text content.
-
-**Key Features**:
-- Recognizes standard section markers: verse, chorus, bridge, pre-chorus, outro, tag, intro
-- Handles Swedish section names: "refräng" → "chorus", "vers" → "verse", "slut" → "outro"
-- Supports numbered sections: "verse 1", "chorus 2"
-- Advanced pattern detection for songs without explicit markers (repeated content detection)
-- Intelligent section boundary detection (doesn't confuse "Chorus content" with "chorus" marker)
-
-**Main Classes**: `SectionDetector`, `AdvancedSectionDetector`
-**Convenience Function**: `detect_sections(text, advanced=False)`
-
-### 3. Text Cleaner (`src/processing/text_cleaner.py`)
-
-**Purpose**: Clean and normalize text after RTF parsing.
-
-**Key Features**:
-- Whitespace normalization (multiple spaces, excessive blank lines)
-- RTF artifact removal (remaining control codes, formatting)
-- Special character replacement (smart quotes, dashes)
-- Song-specific cleaning: chord removal, repetition marker cleaning, capitalization
-- Preserves text structure while cleaning content
-
-**Main Classes**: `TextCleaner`, `SongTextCleaner`
-**Convenience Function**: `clean_text(text, for_song=True, remove_chords=False)`
-
-## Integration with Database Layer
-
-The EasyWorship database module has been enhanced with integrated processing methods:
-
-- `get_song_with_processed_lyrics()`: Complete processing pipeline for a single song
-- `get_all_songs_with_processed_lyrics()`: Batch processing for entire song library
-
-## Testing Suite
-
-Comprehensive test suite (`tests/test_rtf_processing.py`) covering:
-- Swedish Unicode character parsing
-- Section detection accuracy
-- Text cleaning functionality
-- Complete integrated workflow
-- Edge cases and error handling
-
-**Test Results**: All 14 tests passing ✓
-
-## Demo Script
-
-Interactive demo (`rtf_demo.py`) showcasing:
-- RTF parsing with Swedish content
-- Section detection examples
-- Text cleaning demonstrations
-- Advanced pattern detection
-- Database integration (when available)
-
-## Key Technical Achievements
-
-### Swedish Language Support
-- Correct handling of Swedish characters: å, ä, ö
-- Proper Unicode escape sequence decoding
-- Swedish section name mapping to English equivalents
-
-### Robust Parsing
-- Primary parsing with striprtf library
-- Fallback manual parsing for malformed content
-- Graceful error handling and logging
-
-### Intelligent Section Detection
-- Context-aware section marker identification
-- Prevents false positives (e.g., "Chorus content" vs "chorus" marker)
-- Advanced pattern recognition for unmarked sections
-
-### Text Quality
-- Comprehensive cleaning and normalization
-- Song-specific formatting enhancements
-- Preservation of intentional structure
-
-## Performance Characteristics
-
-- Fast processing suitable for large song libraries
-- Memory efficient with streaming-capable design
-- Robust error recovery for corrupted or incomplete data
-
-## Usage Examples
-
-```python
-from processing import parse_rtf, detect_sections, clean_text
-
-# Basic RTF processing
-rtf_content = r"{{\rtf1... Djupt inne i hj\u228?rtat...}"
-parsed = parse_rtf(rtf_content)
-cleaned = clean_text(parsed['plain_text'], for_song=True)
-sections = detect_sections(cleaned)
-
-# Database integration
-from database.easyworship import EasyWorshipDatabase
-db = EasyWorshipDatabase("path/to/database")
-song = db.get_song_with_processed_lyrics(song_id)
-```
-
-## Next Steps
-
-The RTF processing modules are now ready for integration with:
-1. GUI interface for user interaction
-2. ProPresenter export functionality  
-3. Batch processing workflows
-4. Error reporting and logging systems
-
-## Files Modified/Created
-
-- `src/processing/rtf_parser.py` - New RTF parsing module
-- `src/processing/section_detector.py` - New section detection module
-- `src/processing/text_cleaner.py` - New text cleaning module
-- `src/processing/__init__.py` - Updated module exports
-- `src/database/easyworship.py` - Enhanced with processing integration
-- `src/database/__init__.py` - Updated module exports
-- `tests/test_rtf_processing.py` - Comprehensive test suite
-- `rtf_demo.py` - Interactive demonstration script
-- `requirements.txt` - Updated with correct striprtf version
-
-This implementation provides a solid foundation for the EasyWorship to ProPresenter conversion project, with particular attention to Swedish language support and robust error handling.
