+++ build_scripts/upload_release_safe.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 """
-Upload Executable to GitHub Release
+Windows-Safe Upload Executable to GitHub Release
 Simple script to upload pre-built executable to existing GitHub release
 """
 
@@ -9,6 +10,10 @@ import sys
 import subprocess
 import hashlib
 from pathlib import Path
+
+# Fix Windows console encoding issues
+if sys.platform == 'win32':
+    os.environ['PYTHONIOENCODING'] = 'utf-8'
 
 def calculate_sha256(file_path):
     """Calculate SHA256 hash of file"""
@@ -21,13 +26,13 @@ def calculate_sha256(file_path):
 def check_github_cli():
     """Check if GitHub CLI is available"""
     try:
-        result = subprocess.run(['gh', '--version'], 
-                              capture_output=True, text=True, check=True)
-        print("✅ GitHub CLI available")
+        subprocess.run(['gh', '--version'], 
+                      capture_output=True, text=True, check=True)
+        print("[OK] GitHub CLI available")
         return True
     except (subprocess.CalledProcessError, FileNotFoundError):
-        print("❌ GitHub CLI not found")
-        print("📥 Install from: https://cli.github.com/")
+        print("[ERROR] GitHub CLI not found")
+        print("[INFO] Install from: https://cli.github.com/")
         return False
 
 def list_releases():
@@ -35,7 +40,7 @@ def list_releases():
     try:
         result = subprocess.run(['gh', 'release', 'list', '--limit', '10'], 
                               capture_output=True, text=True, check=True)
-        print("📋 Recent releases:")
+        print("[INFO] Recent releases:")
         for line in result.stdout.split('\n')[:5]:  # Show top 5
             if line.strip():
                 parts = line.split('\t')
@@ -45,24 +50,24 @@ def list_releases():
                     print(f"   {tag} ({status})")
         return True
     except subprocess.CalledProcessError:
-        print("❌ Could not list releases")
+        print("[ERROR] Could not list releases")
         return False
 
 def upload_to_release(version, exe_path):
     """Upload executable to GitHub release"""
-    print(f"📤 Uploading to release v{version}...")
+    print(f"[UPLOAD] Uploading to release v{version}...")
     
     if not exe_path.exists():
-        print(f"❌ Executable not found: {exe_path}")
+        print(f"[ERROR] Executable not found: {exe_path}")
         return False
     
     # Calculate hash
     sha256 = calculate_sha256(exe_path)
     size_mb = exe_path.stat().st_size / (1024 * 1024)
     
-    print(f"📁 File: {exe_path}")
-    print(f"📏 Size: {size_mb:.2f} MB") 
-    print(f"🔒 SHA256: {sha256}")
+    print(f"[INFO] File: {exe_path}")
+    print(f"[INFO] Size: {size_mb:.2f} MB") 
+    print(f"[INFO] SHA256: {sha256}")
     
     try:
         # Upload to release
@@ -72,8 +77,8 @@ def upload_to_release(version, exe_path):
             '--clobber'  # Overwrite if exists
         ], check=True)
         
-        print(f"✅ Successfully uploaded to release v{version}")
-        print(f"🌐 View at: https://github.com/karllinder/ewexport/releases/tag/v{version}")
+        print(f"[OK] Successfully uploaded to release v{version}")
+        print(f"[INFO] View at: https://github.com/karllinder/ewexport/releases/tag/v{version}")
         
         # Update release description with hash
         release_body = f"""SHA256: `{sha256}`
@@ -83,28 +88,28 @@ def upload_to_release(version, exe_path):
 If your antivirus flags this file:
 1. Verify the SHA256 hash matches the one above
 2. Add an exception for ewexport.exe
-3. See [ANTIVIRUS.md](https://github.com/karllinder/ewexport/blob/main/ANTIVIRUS.md) for guidance"""
+3. See the project README for antivirus guidance"""
         
-        # Try to update release notes (optional, don't fail if this doesn't work)
+        # Try to update release notes (optional)
         try:
             subprocess.run([
                 'gh', 'release', 'edit', f'v{version}',
-                '--notes-file', '-'
-            ], input=release_body, text=True, check=True)
-            print("📝 Updated release notes with SHA256")
+                '--notes', release_body
+            ], check=True)
+            print("[INFO] Updated release notes with SHA256")
         except:
-            print("⚠️  Could not update release notes (this is OK)")
+            print("[WARNING] Could not update release notes (this is OK)")
         
         return True
         
     except subprocess.CalledProcessError as e:
-        print(f"❌ Upload failed: {e}")
+        print(f"[ERROR] Upload failed: {e}")
         return False
 
 def main():
     """Main upload process"""
     print("=" * 50)
-    print("📤 Upload Executable to GitHub Release")
+    print("Upload Executable to GitHub Release (Windows Safe)")
     print("=" * 50)
     
     # Check prerequisites
@@ -114,8 +119,8 @@ def main():
     # Find executable
     exe_path = Path('dist/ewexport.exe')
     if not exe_path.exists():
-        print(f"❌ Executable not found: {exe_path}")
-        print("💡 Run 'python build_clean.py' first to build it")
+        print(f"[ERROR] Executable not found: {exe_path}")
+        print("[INFO] Run 'python build_scripts/build_clean.py' first to build it")
         return False
     
     # List releases
@@ -124,9 +129,9 @@ def main():
     print()
     
     # Get version to upload to
-    version = input("Enter version to upload to (e.g., 1.1.6): ").strip()
+    version = input("Enter version to upload to (e.g., 1.2.0): ").strip()
     if not version:
-        print("❌ No version specified")
+        print("[ERROR] No version specified")
         return False
     
     # Remove 'v' prefix if provided
@@ -134,21 +139,21 @@ def main():
         version = version[1:]
     
     # Confirm upload
-    print(f"\n📋 Ready to upload:")
+    print(f"\n[INFO] Ready to upload:")
     print(f"   File: {exe_path}")
     print(f"   To: Release v{version}")
     
     response = input("\nProceed with upload? (y/n): ").strip().lower()
     if response != 'y':
-        print("❌ Upload cancelled")
+        print("[CANCELLED] Upload cancelled")
         return False
     
     # Upload
     if upload_to_release(version, exe_path):
-        print("\n🎉 Upload successful!")
+        print("\n[SUCCESS] Upload successful!")
         return True
     else:
-        print("\n❌ Upload failed")
+        print("\n[ERROR] Upload failed")
         return False
 
 if __name__ == "__main__":
@@ -156,8 +161,8 @@ if __name__ == "__main__":
         success = main()
         sys.exit(0 if success else 1)
     except KeyboardInterrupt:
-        print("\n\n❌ Upload cancelled by user")
+        print("\n\n[CANCELLED] Upload cancelled by user")
         sys.exit(1)
     except Exception as e:
-        print(f"\n\n❌ Unexpected error: {e}")
+        print(f"\n\n[ERROR] Unexpected error: {e}")
         sys.exit(1)
